+++ epserde/examples/internal_param_zero.rs
@@ -14,6 +14,7 @@ use epserde::{deser::DeserType, prelude::*, ser::SerType};
 
 #[derive(Epserde, Debug, PartialEq, Eq, Default, Clone)]
 struct Data<A: ZeroCopy> {
+    #[epserde(force_full)]
     a: Vec<A>,
 }